+++ packages/warp-protocol/src/controller/mod.rs
@@ -13,6 +13,9 @@ use crate::controller::template::{
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{Addr, QueryRequest, Uint128, Uint64};
 
+use self::account::WithdrawAssetMsg;
+
+
 pub mod account;
 pub mod condition;
 pub mod job;
@@ -70,6 +73,7 @@ pub enum ExecuteMsg {
     EvictJob(EvictJobMsg),
 
     CreateAccount(CreateAccountMsg),
+    WithdrawAsset(WithdrawAssetMsg),
 
     UpdateConfig(UpdateConfigMsg),
